pyproject.toml
[project]
name = "elec-pokeball3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.5",
    "pygame>=2.6.1",
    "pyglm>=2.8.2",
    "pyopengl>=3.1.9",
]
